Started working on a serial command parser.
Wrote this scratch code to act as a serial command parser for debugging the protocol. I just wrote something fast and terrible that would
work for the time being. It is full of bugs and needs some cleaning up. I haven't tocuhed Python for months. Most bugs are between int and
string conversion. The looping pull from the serial buffer might be a bug as well.

So far the Arduino is able to read the initial "HELLO" but then infinitely sends back garbage after the first handshake. I wonder if the
setup() in controller.cpp is to blame.