Finished serial control on the Arduino side for now.
There are still six commands, but they're more thought out than before. The CONNECTION command was redundant, and passing a signed byte isn't needed for the motor because we can just send REVERSE instead, which may be more useful. STOP was also redundant since stopping is just setting the motor to zero.

Turns out when you use a stepper motor you cannot use AnalogWrite() for pin 10, so the motor pin had to be changed. The motor min is set to 0 on the Arduino side, but the real threshold is governed on the Python side now. I think it makes more sense since a 0 PWN is still valid and technically still works even if it doesn't turn the wheels.