Implemented a working control program.
I decided that writing a debugger was taking too much time! I debugged most of the functionality in IDLE and wrote the enum and methods for control going by what worked. I ran into a good amount of bugs with the buffer that obsfucated functionality, so I decided just to get things working since that's the name of the game right now. I also included a little demo to demonstrate functionality.

Refactoring can come later and optimizationing early is a cardinal sin. At least the car works now!