Python shell working without reflashing Arduino. Will test next.
Major changes:

timeout is now None so we wait until the byte we want.

Input buffer is cleared after connection handshake so all remaining HELLOs are discarded. This could be refactored as it's messy.

Deleted command_parse() as __repr__ and __str__ take care of it.

__repr__ now returns the name of the command plus the value, whereas __str__ returns just the command name.

read_command() now checks if the byte value is in the enum. If so, it returns it. Otherwise it raises an error.