Command parser and debug mode are functioning.
Removed the stop command from reverse(). I'm trusting the motor controller to switch the signal. Reversing is a rare thing anyway, and we once done with debugging we can reenable it. The main reason was to avoid getting back multiple signals from serial.

When exiting, the reset() function leaves commands in the input buffer since it both stops the motor and returns steering to the center. This can either be changed to just stopping on exit (since restarting the program resets the Arduino) or ignored (since the input buffer is cleared after connecting).