Blue lane detection and steering servo angle
Right now there are a lot of magic numbers used for things like the bird's eye view transformation
and image adjustments. Some of these image adjustment numbers are unavoidable as things stand. Maybe
they can be constants later on.

The bird's eye view transformation aids in finding the steering angle, at the very least from a
human standpoint. It's hard to find parallel lines with lens parallax. The resolution can be shrunk
if need be. The lane detection only runs on a small matrix composed of the image's bottom half.
Everything above that is ignored.

Due to the coordinate system in OpenCV images, the steering angle has to be changed to one
compatible with the servo. It actually seems like the servo corresponds to real values. This will be
more apparent after testing.